move util classes into utils and config loader into utils
